test(manifest): re-home G3 suite into per-package unittests (align with #198)
Rebased onto dev/dtni, which adopted the per-package unittests convention
(#198: tests live beside their package, descriptive names, no test_dtni_
prefix). Move the monolithic cvs/lib/unittests/test_dtni_manifest.py into
cvs/lib/manifest/unittests/ split per module:

  - _fixtures.py        shared _full_manifest helper
  - test_schema.py      roundtrip, extra=forbid, B7 no-redaction, NaN/Inf JSON
  - test_layout.py      content-addressable + A1 remote mirror
  - test_events.py      closed vocab, append-only, emit-after-close
  - test_sidecars.py    wide samples / long-format trajectory / resolved config
  - test_export.py      fan-in fact table, empty-tree schema, visible skips

No change to manifest source (schema/layout/events/sidecars/export) or its
A1/B7 behavior; tests-only relocation. All 25 tests pass; ruff clean.
